refactor(playground): move useResizablePanel into its own hooks folder
